Add files via upload
Transform the Cartesian coordinate into polar coordinates. The distance from the current position to the goal position is \rou, the angle between the current pose to the goal vector is \alpha, the angle between the current position in the Golab frame to the goal pose is beta. Then use the PI control law.